==> analysis/2_Run_models.R <==
# when run models make sure make things factors beforehand

# Dependencies
box::use(R / dirs[dirs])
box::use(clean = R / clean_data)
box::use(mod = R / for_analysis)

box::use(readr[read_csv])
box::use(ggplot2[...])
box::use(dplyr[...])
box::use(moments)
box::use(kinship2)
box::use(sf)
box::use(tidyr)
box::use(magrittr[`%>%`])
box::use(asreml[...])
box::use(Matrix)
box::use(brms[...])
# box::use(bayesplot[...])
# box::use(tidybayes[...])
box::use(marginaleffects[...])


# Read in data   ---------------------------------------------------------

gtit_data_sub_nb <-
    read.csv(file.path(dirs$data_output, "GTIT_data_anim_mod.csv"),
        na.strings = c("", "NA")
    ) %>%
    mutate_at(
        c("breeding_year", "mother", "Fbreeding_age_group", "nest_box"),
        as.factor
    )

nrow(gtit_data_sub_nb) # #11658

# to run with a subset
# put dates as range e.g. 2011:2021, or 'ALL' to get all years of data
breed_data <- gtit_data_sub_nb %>%
    mod$choose_data(gtit_data_sub_nb, years = "ALL")
unique(breed_data$breeding_year)
nrow(breed_data) # 11661


# Correlation between lay and hatch date --------------------------------

# get lag between laying and hatching
gtit_data_sub_nb <- gtit_data_sub_nb %>%
    group_by(mother) %>%
    mutate(lay_hatch_lag = april_hatch_date - april_lay_date) %>%
    ungroup()

nrow(subset(gtit_data_sub_nb, !is.na(april_lay_date) &
    !is.na(april_hatch_date))) # 10893

# correlation coefficient r2
cor(gtit_data_sub_nb$april_lay_date, gtit_data_sub_nb$april_hatch_date,
    method = "pearson", use = "complete.obs"
) # 0.952
summary(lm(april_hatch_date ~ april_lay_date, data = gtit_data_sub_nb))
# r2 of 0.906

# and median and sd of lag
median(gtit_data_sub_nb$lay_hatch_lag, na.rm = T) # 21
sd(gtit_data_sub_nb$lay_hatch_lag, na.rm = T) # 2.735


# Prep pedigree ---------------------------------------------------------
# pedigree data input
ped_data <- read.csv(file.path(dirs$data_output, "GTIT-pedigree-for-asreml.csv"), na.strings = c("", "NA"))
nrow(ped_data) # 15744

# get inverse of matrix
ped_data_inv <- asreml::ainverse(ped_data)

# create another column to associate with second matrix

breed_data$mother <- as.factor(breed_data$mother)
breed_data$mother2 <- breed_data$mother


# Read in matrices ---------------------------------------------------------
env_matrix <- readRDS(file.path(dirs$data_output, "env_matrix.rds"))

spatial_matrix <- readRDS(file.path(dirs$data_output, "spatial_matrix.rds"))



# RUN MODELS ---------------------------------------------------------


# LAYING DATE  ------------------------------------------------------------

# model basic
LD_basic <- asreml(
    fixed = april_lay_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        breeding_year,
    data = breed_data
)
saveRDS(LD_basic, file = file.path(dirs$model_output, "LD_basic_output.rds"))


# model with NB
LD_NB <- asreml::asreml(
    fixed = april_lay_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        breeding_year +
        nest_box,
    data = breed_data
)
saveRDS(LD_NB, file = file.path(dirs$model_output, "LD_NB_output.rds"))


# without distance matrix
LD_spat <- asreml::asreml(
    fixed = april_lay_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        vm(mother2, spatial_matrix) +
        breeding_year,
    data = breed_data,
    workspace = 300e6, pworkspace = 300e6
)
saveRDS(LD_spat, file = file.path(dirs$model_output, "LD_spat_output.rds"))



# model with envir sim matrix
LD_envir <- asreml::asreml(
    fixed = april_lay_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        vm(mother2, env_matrix) +
        breeding_year,
    data = breed_data,
    workspace = 300e6, pworkspace = 300e6
)
saveRDS(LD_envir, file = file.path(dirs$model_output, "LD_envir_output.rds"))



# HATCHING DATE -----------------------------------------------------------


# model basic
HD_basic <- asreml::asreml(
    fixed = april_hatch_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        breeding_year,
    data = breed_data
)
saveRDS(HD_basic, file = file.path(dirs$model_output, "HD_basic_output.rds"))


# model with NB
HD_NB <- asreml::asreml(
    fixed = april_hatch_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        breeding_year +
        nest_box,
    data = breed_data
)
saveRDS(HD_NB, file = file.path(dirs$model_output, "HD_NB_output.rds"))


# without distance matrix
HD_spat <- asreml::asreml(
    fixed = april_hatch_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        vm(mother2, spatial_matrix) +
        breeding_year,
    data = breed_data,
    workspace = 300e6, pworkspace = 300e6,
    maxiter = 50
)
saveRDS(HD_spat, file = file.path(dirs$model_output, "HD_dist_output.rds"))



# model with envir sim matrix
HD_envir <- asreml::asreml(
    fixed = april_hatch_date ~ 1 + Fbreeding_age_group,
    random = ~ vm(mother, ped_data_inv) +
        ide(mother, ped_data_inv) +
        vm(mother2, env_matrix) +
        breeding_year,
    data = breed_data,
    workspace = 300e6, pworkspace = 300e6,
    maxiter = 50
)
saveRDS(HD_envir, file = file.path(dirs$model_output, "HD_envir_output.rds"))


# FIXED EFFECTS ---------------------------------------------------------

# get fixed effects of each model in a table

fixed_effects <- rbind(
    get_fixed_effects(LD_basic),
    get_fixed_effects(LD_NB),
    get_fixed_effects(LD_spat),
    get_fixed_effects(LD_envir),
    get_fixed_effects(HD_basic),
    get_fixed_effects(HD_NB),
    get_fixed_effects(HD_spat),
    get_fixed_effects(HD_envir)
)
write.csv(fixed_effects,
    file = file.path(dirs$model_output, "fixed_effects_table.csv"),
    row.names = F
)


# Wald tests ---------------------------------------------------------

# wald tests # P-value for fixed effects
# laying date
asreml::wald.asreml(LD_basic, ssType = "conditional", denDF = "numeric")
asreml::wald.asreml(LD_NB, ssType = "conditional", denDF = "numeric")
asreml::wald.asreml(LD_spat, ssType = "conditional", denDF = "numeric")
asreml::wald.asreml(LD_envir, ssType = "conditional", denDF = "numeric")

# hatching date
asreml::wald.asreml(HD_basic, ssType = "conditional", denDF = "numeric")
asreml::wald.asreml(HD_NB, ssType = "conditional", denDF = "numeric")
asreml::wald.asreml(HD_spat, ssType = "conditional", denDF = "numeric")
asreml::wald.asreml(HD_envir, ssType = "conditional", denDF = "numeric")

# run wald.asreml for each model and save each output as rds file
run_wald(LD_basic, "LD_basic")
run_wald(LD_NB, "LD_NB")
run_wald(LD_spat, "LD_spat")
run_wald(LD_envir, "LD_envir")
run_wald(HD_basic, "HD_basic")
run_wald(HD_NB, "HD_NB")
run_wald(HD_spat, "HD_spat")
run_wald(HD_envir, "HD_envir")
