==> analysis/3.1_Regression_analysis.R <==
# Dependencies
box::use(. / R / dirs[dirs])
box::use(readr[read_csv])
box::use(ggplot2[...])
box::use(dplyr)
box::use(tidyr)
box::use(magrittr[`%>%`])


# Supplementary for paper - van der Juegd and McCleery 2002 redo

# GET MEAN LD OF INDIVS ---------------------------------------------------

# input breeding data - created in 1_Prep_data.R
breed_gtit_sub <- read.csv(file.path(dirs$data_output, "GTIT_data_anim_mod.csv"))

nrow(breed_gtit_sub) # 11658

summary(as.factor(breed_gtit_sub$fem_breed_age))

# regress with year and age at breeding to get residuals
breed_gtit_sub <- breed_gtit_sub %>%
  dplyr::mutate(resid_ld = lm(april_lay_date ~ breeding_year + fem_breed_age)$residuals)

# how many have multiple recorded lay dates
num_lay <- breed_gtit_sub %>%
  dplyr::group_by(., mother) %>%
  dplyr::summarise(., num_ld = length(resid_ld))
# how many more than 1
nrow(subset(num_lay, num_ld > 1)) # 2589

# get mean over breeding attempts
breed_gtit_mothLD <- breed_gtit_sub %>%
  dplyr::group_by(mother) %>%
  dplyr::summarise(resid_ld_mean = mean(resid_ld))

# then can merge these with pedigree
# in pedigree - get mean LD of id, of mother and of Father
ped_gtit <- read.csv(file.path(dirs$data_output, "GTIT-pedigree-for-asreml.csv"))

nrow(breed_gtit_mothLD) # 7679
nrow(ped_gtit) # 15744

# merge - to get ID mean LD for id and dam
ped_gtit_LD <- dplyr::inner_join(ped_gtit, breed_gtit_mothLD, by = c("id" = "mother")) %>%
  dplyr::rename(resid_ld_mean_id = resid_ld_mean) %>%
  dplyr::select(id, dam, sire, resid_ld_mean_id) %>%
  dplyr::inner_join(., breed_gtit_mothLD, by = c("dam" = "mother")) %>%
  dplyr::rename(resid_ld_mean_dam = resid_ld_mean) %>%
  dplyr::select(id, dam, sire, resid_ld_mean_id, resid_ld_mean_dam)

nrow(ped_gtit_LD) # 3378

# MERGE TO MAKE DATA WITH LD, YR, DISP DIST  ------------------------------------------------------

# use pedigree - created in script 1_Prep_data
ped_big <- read.csv(file.path(dirs$data_output, "GTIT-pedigree-full.csv"))
head(ped_big)
# get coordinates from nestbox data - from raw dat - keep 1 row for each box
habitat_data <- read.csv(file.path(dirs$data_output, "Habitat_data_nestboxes.csv")) %>%
  dplyr::distinct(box, .keep_all = T) %>%
  dplyr::select(box, x, y)
nrow(habitat_data) # 1016

# get box for mothers first breeding attempt as already have where born
disp_LD <- breed_gtit_sub %>%
  dplyr::group_by(mother) %>%
  dplyr::slice_min(order_by = breeding_year) %>%
  dplyr::ungroup() %>%
  dplyr::select(mother, nest_box) %>%
  dplyr::inner_join(ped_big, ., by = c("bto_ring" = "mother")) %>%
  dplyr::rename(
    nestbox_firstbr = nest_box,
    id = bto_ring,
    dam = mother
  ) %>%
  dplyr::select(id, dam, nest_box_born_in, nestbox_firstbr) %>%
  # merge with nestbox to get locations and distance
  dplyr::inner_join(., habitat_data, by = c("nest_box_born_in" = "box")) %>%
  dplyr::rename(
    x_born = x,
    y_born = y
  ) %>%
  dplyr::inner_join(., habitat_data, by = c("nestbox_firstbr" = "box")) %>%
  dplyr::rename(
    x_breed = x,
    y_breed = y
  ) %>%
  dplyr::mutate(disp_dist = sqrt((x_breed - x_born)^2 + (y_breed - y_born)^2)) %>%
  # merge to get LDs
  dplyr::inner_join(., ped_gtit_LD, by = c("id", "dam"))

# nest_box_born in is bto_ring's nest box born in
# nestbox_firstbr is bto_ring's first nest box they bred in (dispersal box)

nrow(disp_LD) # 3371
# get number of unique birds in id and dam column together
length(unique(c(disp_LD$id, disp_LD$dam))) # 3191

median(disp_LD$disp_dist)
mean(disp_LD$disp_dist)

# percentage of bird staying within 1000m
nrow(subset(disp_LD, disp_dist <= 1000)) / nrow(disp_LD) # 0.75

saveRDS(disp_LD, file = file.path(dirs$data_output, "mothers_daughter_disp_LD.rds"))


# PLOT OF DISPERSAL DISTANCE


# just females with lines on for splitting up distance
natal_disp_plot <-
  ggplot(
    subset(disp_LD),
    aes(x = disp_dist)
  ) +
  geom_histogram(bins = 50, colour = "white", fill = "cornflower blue") +
  xlab("Natal dispersal distance, metres") +
  ylab("Number of individuals") +
  geom_vline(xintercept = 500, linewidth = 1, color = "black") +
  geom_vline(xintercept = 1000, linewidth = 1, color = "black") +
  theme(
    panel.background = element_rect(fill = "white", color = "white"),
    plot.background = element_rect(fill = "white", color = "white"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.y = element_line(color = "grey"),
    # axis.title.x = element_text(colour = "#a4948c"),
    # axis.title.y = element_text(colour = "#a4948c"),
    text = element_text(color = "black", face = "bold", size = 16),
    axis.text = element_text(color = "black", face = "bold", size = 12)
  )

ggsave(
  filename = "./plots/natal_dispersal_distance.pdf",
  plot = natal_disp_plot,
  scale = 1,
  width = 7,
  height = 15,
  units = "cm",
  dpi = 350,
  limitsize = FALSE,
  bg = "transparent"
)




# PLOTS OF LAY DATES

# what is heritability?
# heritabiltiy = slope * 2
summary(lm(disp_LD$resid_ld_mean_id ~ disp_LD$resid_ld_mean_dam))$coefficient[2] * 2 # 0.23


# GROUPING DISPERSAL DISTANCE

# make column classing disp distance to short, medium, large

disp_LD <- disp_LD %>%
  dplyr::mutate(
    disp_dist_split =
      dplyr::case_when(
        disp_dist <= 500 ~ "short",
        disp_dist > 500 & disp_dist <= 1000 ~ "medium",
        disp_dist > 1000 ~ "long"
      )
  )
table(disp_LD$disp_dist_split)

# mean of each group
disp_LD_summ <- disp_LD %>%
  dplyr::group_by(., disp_dist_split) %>%
  dplyr::summarise(.,
    length = length(disp_dist_split),
    mean_disp = mean(disp_dist),
    median_disp = median(disp_dist)
  )
disp_LD_summ


# linear models
short_output <- summary(lm(resid_ld_mean_id ~ resid_ld_mean_dam,
  data = subset(disp_LD, disp_dist_split == "short")
))
# 0.17986 +- 0.03712, p < 0.001, N = 962
short_output$coefficient[2] * 2 # 0.37

medium_output <- summary(lm(resid_ld_mean_id ~ resid_ld_mean_dam,
  data = subset(disp_LD, disp_dist_split == "medium")
))
# 0.09436 +- 0.03584, p = 0.009, N = 1075
medium_output$coefficient[2] * 2 # 0.19

long_output <- summary(lm(resid_ld_mean_id ~ resid_ld_mean_dam,
  data = subset(disp_LD, disp_dist_split == "long")
))
# 0.07417 +- 0.03227, p = 0.022, N = 1250
long_output$coefficient[2] * 2 # 0.15


# overall regression plot
LD_regression_plot_overall <-
  ggplot(
    disp_LD,
    aes(x = resid_ld_mean_dam, y = resid_ld_mean_id)
  ) +
  geom_point(alpha = 0.4, colour = "cornflower blue") +
  geom_smooth(method = "lm", colour = "#3159A0") +
  xlab("Mother's mean laying date") +
  ylab("Daughter's mean laying date") +
  theme(
    panel.background = element_rect(fill = "white", color = "white"),
    plot.background = element_rect(fill = "white", color = "white"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.y = element_line(color = "grey"),
    # axis.title.x = element_text(colour = "#a4948c"),
    # axis.title.y = element_text(colour = "#a4948c"),
    text = element_text(color = "black", face = "bold", size = 16),
    axis.text = element_text(color = "black", face = "bold", size = 12),
    legend.position = c(.972, .972),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6),
    legend.background = element_blank(),
    legend.box.background = element_blank(),
    strip.text.x = element_text(size = 14, colour = "black", face = "bold"),
    strip.background = element_rect(colour = "white", fill = "white")
  )

ggsave(
  filename = "./plots/LD_regression_plot_overall.pdf",
  plot = LD_regression_plot_overall,
  scale = 1,
  width = 15,
  height = 10,
  units = "cm",
  dpi = 350,
  limitsize = FALSE,
  bg = "transparent"
)

# plot with facet
LD_regression_plot_facet <-
  ggplot(
    transform(disp_LD,
      disp_dist_split = factor(disp_dist_split,
        levels = c("short", "medium", "long")
      )
    ),
    aes(x = resid_ld_mean_dam, y = resid_ld_mean_id)
  ) +
  geom_point(alpha = 0.4, colour = "cornflower blue") +
  geom_smooth(method = "lm", colour = "#3159A0") +
  facet_wrap(~disp_dist_split) +
  xlab("Mother's mean laying date") +
  ylab("Daughter's mean laying date") +
  theme(
    panel.background = element_rect(fill = "white", color = "white"),
    plot.background = element_rect(fill = "white", color = "white"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.y = element_line(color = "grey"),
    # axis.title.x = element_text(colour = "#a4948c"),
    # axis.title.y = element_text(colour = "#a4948c"),
    text = element_text(color = "black", face = "bold", size = 16),
    axis.text = element_text(color = "black", face = "bold", size = 12),
    legend.position = c(.972, .972),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6),
    legend.background = element_blank(),
    legend.box.background = element_blank(),
    strip.text.x = element_text(size = 14, colour = "black", face = "bold"),
    strip.background = element_rect(colour = "white", fill = "white")
  )

ggsave(
  filename = "./plots/LD_regression_plot_facet.pdf",
  plot = LD_regression_plot_facet,
  scale = 1,
  width = 45,
  height = 10,
  units = "cm",
  dpi = 350,
  limitsize = FALSE,
  bg = "transparent"
)


# HATCHING DATE


# remove those with no hatch date
breed_gtit_sub_hat <- breed_gtit_sub %>%
  dplyr::filter(!is.na(april_hatch_date))
nrow(breed_gtit_sub_hat) # 10890

# regress with year and age at breeding to get residuals
breed_gtit_sub_hat <- breed_gtit_sub_hat %>%
  dplyr::mutate(resid_hd = lm(april_hatch_date ~ breeding_year + fem_breed_age)$residuals)

# how many have multiple recorded lay dates
num_hat <- breed_gtit_sub_hat %>%
  dplyr::group_by(., mother) %>%
  dplyr::summarise(., num_hd = length(resid_hd))
# how many more than 1
nrow(subset(num_hat, num_hd > 1)) # 2383

# get mean over breeding attempts
breed_gtit_mothHD <- breed_gtit_sub_hat %>%
  dplyr::group_by(mother) %>%
  dplyr::summarise(resid_hd_mean = mean(resid_hd))

# merge - to get ID mean LD for id and dam
ped_gtit_HD <- dplyr::inner_join(ped_gtit, breed_gtit_mothHD, by = c("id" = "mother")) %>%
  dplyr::rename(resid_hd_mean_id = resid_hd_mean) %>%
  dplyr::select(id, dam, sire, resid_hd_mean_id) %>%
  dplyr::inner_join(., breed_gtit_mothHD, by = c("dam" = "mother")) %>%
  dplyr::rename(resid_hd_mean_dam = resid_hd_mean) %>%
  dplyr::select(id, dam, sire, resid_hd_mean_id, resid_hd_mean_dam)

head(ped_gtit_HD)
nrow(ped_gtit_HD) # 3191

# MERGE TO MAKE DATA WITH LD, YR, DISP DIST  ------------------------------------------------------

# get mothers first breeding attempt as already have where born
disp_HD <- breed_gtit_sub_hat %>%
  dplyr::group_by(mother) %>%
  dplyr::slice_min(order_by = breeding_year) %>%
  dplyr::ungroup() %>%
  dplyr::select(mother, nest_box) %>%
  dplyr::inner_join(ped_big, ., by = c("bto_ring" = "mother")) %>%
  dplyr::rename(
    nestbox_firstbr = nest_box,
    id = bto_ring,
    dam = mother
  ) %>%
  dplyr::select(id, dam, nest_box_born_in, nestbox_firstbr) %>%
  # merge with nestbox to get locations and distance
  dplyr::inner_join(., habitat_data, by = c("nest_box_born_in" = "box")) %>%
  dplyr::rename(
    x_born = x,
    y_born = y
  ) %>%
  dplyr::inner_join(., habitat_data, by = c("nestbox_firstbr" = "box")) %>%
  dplyr::rename(
    x_breed = x,
    y_breed = y
  ) %>%
  dplyr::mutate(disp_dist = sqrt((x_breed - x_born)^2 + (y_breed - y_born)^2)) %>%
  # merge to get LDs
  dplyr::inner_join(., ped_gtit_HD, by = c("id", "dam"))

# nest_box_born in is bto_ring's nest box born in
# nestbox_firstbr is bto_ring's first nest box they bred in (dispersal box)

head(disp_HD)
nrow(disp_HD) # 3187

median(disp_HD$disp_dist)
mean(disp_HD$disp_dist)

saveRDS(disp_HD, file = file.path(dirs$data, "mothers_daughter_disp_HD.rds"))



# PLOTS OF LAY DATES

# heritabiltiy = slope * 2
summary(lm(disp_HD$resid_hd_mean_id ~ disp_HD$resid_hd_mean_dam))$coefficient[2] * 2 # 0.24



# GROUPING DISPERSAL DISTANCE

# make column classing disp distance to short, medium, large

disp_HD <- disp_HD %>%
  dplyr::mutate(
    disp_dist_split =
      dplyr::case_when(
        disp_dist <= 500 ~ "short",
        disp_dist > 500 & disp_dist <= 1000 ~ "medium",
        disp_dist > 1000 ~ "long"
      )
  )
table(disp_HD$disp_dist_split)

# mean of each group
disp_HD_summ <- disp_HD %>%
  dplyr::group_by(., disp_dist_split) %>%
  dplyr::summarise(.,
    length = length(disp_dist_split),
    mean_disp = mean(disp_dist),
    median_disp = median(disp_dist)
  )
disp_HD_summ



# linear models
short_output_hd <- summary(lm(resid_hd_mean_id ~ resid_hd_mean_dam,
  data = subset(disp_HD, disp_dist_split == "short")
))
# 0.17986 +- 0.03712, p < 0.001, N = 962
short_output_hd$coefficient[2] * 2

medium_output_hd <- summary(lm(resid_hd_mean_id ~ resid_hd_mean_dam,
  data = subset(disp_HD, disp_dist_split == "medium")
))
# 0.09436 +- 0.03584, p = 0.009, N = 1075
medium_output_hd$coefficient[2] * 2


long_output_hd <- summary(lm(resid_hd_mean_id ~ resid_hd_mean_dam,
  data = subset(disp_HD, disp_dist_split == "long")
))
# 0.07417 +- 0.03227, p = 0.022, N = 1250
long_output_hd$coefficient[2] * 2



# overall regression plot
HD_regression_plot_overall <-
  ggplot(
    disp_HD,
    aes(x = resid_hd_mean_dam, y = resid_hd_mean_id)
  ) +
  geom_point(alpha = 0.4, colour = "cornflower blue") +
  geom_smooth(method = "lm", colour = "#3159A0") +
  xlab("Mother's mean hatching date") +
  ylab("Daughter's mean hatching date") +
  theme(
    panel.background = element_rect(fill = "white", color = "white"),
    plot.background = element_rect(fill = "white", color = "white"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.y = element_line(color = "grey"),
    # axis.title.x = element_text(colour = "#a4948c"),
    # axis.title.y = element_text(colour = "#a4948c"),
    text = element_text(color = "black", face = "bold", size = 16),
    axis.text = element_text(color = "black", face = "bold", size = 12),
    legend.position = c(.972, .972),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6),
    legend.background = element_blank(),
    legend.box.background = element_blank(),
    strip.text.x = element_text(size = 14, colour = "black", face = "bold"),
    strip.background = element_rect(colour = "white", fill = "white")
  )

ggsave(
  filename = "./plots/HD_regression_plot_overall.pdf",
  plot = HD_regression_plot_overall,
  scale = 1,
  width = 15,
  height = 10,
  units = "cm",
  dpi = 350,
  limitsize = FALSE,
  bg = "transparent"
)

# plot with facet - for write up!
HD_regression_plot_facet <- ggplot(
  transform(disp_HD,
    disp_dist_split = factor(disp_dist_split,
      levels = c("short", "medium", "long")
    )
  ),
  aes(x = resid_hd_mean_dam, y = resid_hd_mean_id)
) +
  geom_point(alpha = 0.4, colour = "aquamarine3") +
  geom_smooth(method = "lm", colour = "#29705c") +
  facet_wrap(~disp_dist_split) +
  xlab("mother's mean hatching date") +
  ylab("Daughter's mean hatching date") +
  theme(
    panel.background = element_rect(fill = "white", color = "white"),
    plot.background = element_rect(fill = "white", color = "white"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.y = element_line(color = "grey"),
    # axis.title.x = element_text(colour = "#a4948c"),
    # axis.title.y = element_text(colour = "#a4948c"),
    text = element_text(color = "black", face = "bold", size = 16),
    axis.text = element_text(color = "black", face = "bold", size = 12),
    legend.position = c(.972, .972),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6),
    legend.background = element_blank(),
    legend.box.background = element_blank(),
    strip.text.x = element_text(size = 14, colour = "black", face = "bold"),
    strip.background = element_rect(colour = "white", fill = "white")
  )

ggsave(
  filename = "./plots/HD_regression_plot_facet.pdf",
  plot = HD_regression_plot_facet,
  scale = 1,
  width = 15,
  height = 10,
  units = "cm",
  dpi = 350,
  limitsize = FALSE,
  bg = "transparent"
)
